Extract VirtAddrAllocator and add alloc_specific API
